requirements.txt
ollama>=0.1.0
openwebui>=0.1.0
langchain>=0.1.0
llama-cpp-python>=0.2.0
redis>=5.0.0
sqlalchemy>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0
watchdog>=3.0.0
flask>=2.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
pydantic>=2.5.0
pyyaml>=6.0.0
cryptography>=41.0.0
gitpython>=3.1.40
diffusers>=0.23.0
